Split SharedSqlDriver
