+++ cypher-to-sql/src/test/scala/ingraph/compiler/sql/driver/SharedSqlDriver.scala
@@ -1,44 +1,9 @@
 package ingraph.compiler.sql.driver
 
-import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, Suite}
+import org.scalatest.Suite
 
-trait SharedSqlDriver extends BeforeAndAfterAll with BeforeAndAfterEach {
+trait SharedSqlDriver extends SharedDriver[SqlDriver, SqlSession] {
   this: Suite =>
 
-  var driver: SqlDriver = _
-  var session: SqlSession = _
-
-  def initNewDriver(): SqlDriver = new SqlDriver(translateCreateQueries = true)
-
-  protected def init(): Unit = {
-    driver = initNewDriver()
-    session = driver.session()
-  }
-
-  override protected def beforeAll(): Unit = {
-    super.beforeAll()
-
-    init()
-  }
-
-  override protected def beforeEach(): Unit = {
-    super.beforeEach()
-
-    if (!session.isOpen)
-      init()
-  }
-
-  override protected def afterEach(): Unit = {
-    if (!session.isOpen)
-      driver.close()
-
-    super.afterEach()
-  }
-
-  override protected def afterAll(): Unit = {
-    session.close()
-    driver.close()
-
-    super.afterAll()
-  }
+  override def initNewDriver(): SqlDriver = new SqlDriver(translateCreateQueries = true)
 }
